Parse effect specs
parser.mly isn't built with dune, but with additional Makefile targets.
See [1] for a starting point. Here we use only the essential Menhir
invocation.

We're also not concerned with bootstrapping or versioning, so we
generate the entire contents of the boot directory (also making this
project agnostic to the version of Menhir installed).

[1]: https://github.com/ocaml/ocaml/blob/trunk/parsing/HACKING.adoc