Fix the finite loop problem
The problem is inside `simplify_pure`, which is quite surprising.
The problem is that, `simplify_pure` will keep simplify our formula
until convergence, which is denoted by a boolean. There is a mistake
somewhere with the boolean, such that the loop never exits.

The current fix is to bound the loop with another parameter "fuel",
which is decreased by 1 after each iteration. Once the fuel is exhausted
we will stop the loop (this is a technique that is used in Coq to write
non-structurally recursive function).

By the way, the way we write "simplify_term" and "simplify_pure" is too
messy, but it is not our top priority now to fix that.